pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=3.0.0",
    "django>=5.2.2",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
    "rollbar>=1.3.0",
    "whitenoise>=6.9.0",
]
